Fixed styling of product details
