+++ client/src/index.css
@@ -40,12 +40,13 @@ code {
 
 .left-section , .right-section {
   flex-basis: 100px;
-}
+} */
 
 .product-img {
-  height: 600px;
-  width: 400px;
-} */
+  height: 435px;
+  width: 435px;
+  display: block;
+}
 
 /* .right-section {
   float: right;
